nsiproxy: Don't use long types in internal structure / helpers.

+++ dlls/nsiproxy.sys/nsiproxy_private.h
@@ -30,7 +30,7 @@ struct icmp_send_echo_params
 {
     SOCKADDR_INET *dst;
     void *request, *reply;
-    DWORD request_size, reply_len;
+    UINT request_size, reply_len;
     BYTE bits, ttl, tos;
     HANDLE handle;
 };
@@ -38,27 +38,27 @@ struct icmp_send_echo_params
 /* output for IOCTL_NSIPROXY_WINE_ICMP_ECHO - cf. ICMP_ECHO_REPLY */
 struct icmp_echo_reply_32
 {
-    ULONG addr;
-    ULONG status;
-    ULONG round_trip_time;
+    UINT addr;
+    UINT status;
+    UINT round_trip_time;
     USHORT data_size;
     USHORT num_of_pkts;
-    ULONG data_ptr;
+    UINT data_ptr;
     struct
     {
         BYTE ttl;
         BYTE tos;
         BYTE flags;
         BYTE options_size;
-        ULONG options_ptr;
+        UINT options_ptr;
     } opts;
 };
 
 struct icmp_echo_reply_64
 {
-    ULONG addr;
-    ULONG status;
-    ULONG round_trip_time;
+    UINT addr;
+    UINT status;
+    UINT round_trip_time;
     USHORT data_size;
     USHORT num_of_pkts;
     ULONGLONG data_ptr;
